consolidate parser, pipeline, crawler, and spider;
